+++ starfish/test/test_stack_intensity_table_reshaping.py
@@ -13,7 +13,7 @@ def test_reshaping_between_stack_and_intensities():
     the created Imagestack is the same as the original
     """
     np.random.seed(777)
-    image = ImageStack.from_numpy_array(np.random.rand(1, 2, 3, 4, 5))
+    image = ImageStack.from_numpy_array(np.random.rand(1, 2, 3, 4, 5).astype(np.float32))
     pixel_intensities = IntensityTable.from_image_stack(image, 0, 0, 0)
     image_shape = (image.shape['z'], image.shape['y'], image.shape['x'])
     image_from_pixels = pixel_intensities_to_imagestack(pixel_intensities, image_shape)
